+++ src/main/java/gift/client/KakaoApiClient.java
@@ -3,13 +3,13 @@ package gift.client;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gift.config.properties.KakaoProperties;
+import gift.dto.giftorder.GiftOrderResponse;
 import gift.dto.kakao.KakaoAuthResponse;
 import gift.dto.kakao.KakaoTokenResponse;
 import gift.dto.kakao.template.KakaoTemplate;
 import gift.dto.kakao.template.KakaoTemplateCommerce;
 import gift.dto.kakao.template.KakaoTemplateContent;
 import gift.dto.kakao.template.KakaoTemplateLink;
-import gift.dto.order.GiftOrderResponse;
 import gift.exception.BadRequestException;
 import gift.exception.InvalidKakaoTokenException;
 import org.springframework.http.HttpStatus;
