+++ src/test/java/gift/service/auth/AuthServiceTest.java
@@ -3,9 +3,9 @@ package gift.service.auth;
 import gift.config.properties.JwtProperties;
 import gift.dto.auth.LoginRequest;
 import gift.dto.auth.RegisterRequest;
+import gift.dto.giftorder.GiftOrderResponse;
 import gift.dto.kakao.KakaoAuthInformation;
 import gift.dto.kakao.KakaoTokenResponse;
-import gift.dto.order.GiftOrderResponse;
 import gift.exception.DuplicatedEmailException;
 import gift.exception.InvalidLoginInfoException;
 import gift.model.Member;
@@ -14,6 +14,7 @@ import gift.model.OauthToken;
 import gift.model.OauthType;
 import gift.reflection.AuthTestReflectionComponent;
 import gift.repository.MemberRepository;
+import gift.service.KakaoService;
 import gift.service.MemberService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
